requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pypdf
sentence-transformers
numpy
faiss-cpu
